+++ Entities/HeldDevice.h
@@ -281,6 +281,42 @@ ClassInfoGetters
 
 	void SetSupportOffset(Vector newOffset) { m_SupportOffset = newOffset; }
 
+    /// <summary>
+    /// Gets whether this HeldDevice has any limitations on who can pick it up.
+    /// </summary>
+    /// <returns>Whether this HeldDevice has any limitations on who can pick it up.</returns>
+    bool HasPickupLimitations() const { return !m_ActorsWhoCanPickThisUp.empty(); }
+
+    /// <summary>
+    /// Gets whether no Actors (or child classes) can pick up this HeldDevice.
+    /// </summary>
+    /// <returns>Whether no Actors (or child classes) can pick up this HeldDevice.</returns>
+    bool GetNoActorsCanPickThisUp() const { return m_ActorsWhoCanPickThisUp.find("NONE") != m_ActorsWhoCanPickThisUp.end(); }
+
+    /// <summary>
+    /// Sets whether no Actors (or child classes) can pick up this HeldDevice.
+    /// </summary>
+    /// <param name="noActors">Whether no Actors (or child classes) should be able to pick up this HeldDevice. True means none will be able to, false means any other limitations will apply normally.</param>
+    void SetNoActorsCanPickThisUp(bool noActors);
+
+    /// <summary>
+    /// Checks whether Actors with the given PresetName can pick up this HeldDevice.
+    /// </summary>
+    /// <param name="actor">The PresetName of the Actor to check. Ownership is NOT transferred.</param>
+    /// <returns>Whether Actors with the given PresetName can pick up this HeldDevice.</returns>
+    bool ActorCanPickThisUp(const std::string &actorPresetName) { return m_ActorsWhoCanPickThisUp.find(actorPresetName) != m_ActorsWhoCanPickThisUp.end(); }
+
+    /// <summary>
+    /// Allow any Actors (or child classes) with the given PresetName to pick up this HeldDevice, as long as it's not set so no Actors can pick it up.
+    /// </summary>
+    /// <param name="actorPresetName">The PresetName of an Actor who should be able to pick up this HeldDevice.</param>
+    void AddActorWhoCanPickThisUp(const std::string &actorPresetName) { m_ActorsWhoCanPickThisUp.insert(actorPresetName); }
+
+    /// <summary>
+    /// Remove allowance for any Actors (or child classes) with the given PresetName from picking up this HeldDevice. Note that if no specific Actors are allowed to pick this HeldDevice up, all Actors will be allowed to do so unless it's set so no Actors can pick it up.
+    /// </summary>
+    /// <param name="actorPresetName">The PresetName of an Actor who should no longer be able to pick up this HeldDevice.</param>
+    void RemoveActorWhoCanPickThisUp(const std::string &actorPresetName) { m_ActorsWhoCanPickThisUp.erase(m_ActorsWhoCanPickThisUp.find(actorPresetName)); }
 
     /// <summary>
     /// Gets the multiplier for how well this HeldDevice can be gripped by Arms.
@@ -589,6 +625,7 @@ protected:
     float m_MaxSharpLength;
     // If this HeldDevice is currently being supported by a second hand.
     bool m_Supported;
+    std::unordered_set<std::string> m_ActorsWhoCanPickThisUp; //!< The unordered set of Actors (and appropriate subclasses) who can pick up this held device if it's dropped. An empty set means any Actors can pick it up, a "NONE" entry means none can pick it up.
     float m_GripStrengthMultiplier; //!< The multiplier for how well this HeldDevice can be gripped by Arms.
     // Blink timer for the icon
     Timer m_BlinkTimer;